Added test
A good example of where unit tests truly became useful is the labelCase
filter you wrote.

Say someone applied that filter to an element that's model was an
array, or some other case in which .toUpperCase()/.toLowerCase would
not behave well,

We can write additional tests beyond happy path, so we can confirm as
time goes on, and as the functionality of that filter expands, that it
behaves consistently in edge cases.